requirements.txt
langchain-community
langchain-openai
openai
python-dotenv
faiss-cpu
pydantic
streamlit
unstructured
crewai
crewai-tools
psycopg2-binary
embedchain[postgres]
fastapi
uvicorn[standard]
boto3